[Frontend][Admin] Quản lí tác giả, nhà xuất bản

+++ react-frontend/src/admin/components/author/datatableauthor/Datatable.jsx
@@ -1,37 +1,71 @@
-import "./datatable.scss";
 import { DataGrid } from "@mui/x-data-grid";
-import { Link } from "react-router-dom";
 import { useEffect, useState } from "react";
 import BookService from "../../../../service/BookService";
-import NewAuthor from "../newauthor/NewAuthor";
-import DetailAuthor from "../detailauthor/Detailauthor";
+import AuthorModal from "../../author/newauthor/NewAuthor";
+import Button from '@mui/material/Button';
+import "./datatable.scss";
 
 const DatatableAuthor = () => {
+
   const [authors, setAuthors] = useState([]);
-  const [author, setAuthor] = useState([])
-  const [openmodaladd, setOpenmodalAdd] = useState(false);
-  const [openmodalupdate, setOpenmodalUpdate] = useState(false);
 
-  const handleModalAdd = () => {
-    setOpenmodalAdd(true);
-  };
+  const [open, setOpen] = useState(false)
+  const [formData, setFormData] = useState({ id: '', firstName: '', lastName: '', companyName:'' })
 
-  
-  const handleModalView = async (id) => {
-    setOpenmodalUpdate(true);
-    console.log(openmodalupdate)
-    // console.log(id)
-    await BookService.getAuthorbyId(id)
-    .then((res) => {
-      setAuthor(res.data)
-      console.log(res.data);
-    })
-  };
+  const handleClickOpen = () => {
+    setOpen(true);
+  }
 
-  const handleDelete =  async (id) => {
-    await BookService.deleteAuthorById(id);
-    setAuthors(authors.filter(item => item.id != id));
-  };
+  const handleClose = () => {
+    setOpen(false);
+    setFormData({ id: '', firstName: '', lastName: '', companyName: '' });
+  }
+
+  const onChange = (e) => {
+    setFormData({ ...formData, [e.target.name]: e.target.value })
+  }
+
+
+  const handleFormSubmit = async () => {
+
+    try {
+      const id = parseInt(formData.id);
+      if (id) {
+        const updateAuthor = {
+          firstName: formData.firstName,
+          lastName: formData.lastName,
+          companyName: formData.companyName
+        }
+        const respone = await BookService.putAuthorById(id, updateAuthor)
+        handleClose();
+        window.location.reload();
+      } else {
+        const newAuthor = {
+          firstName: formData.firstName,
+          lastName: formData.lastName,
+          companyName: formData.companyName
+        }
+        const response = await BookService.postNewAuthor(newAuthor)
+        handleClose();
+        window.location.reload();
+      }
+    } catch (error) {
+      alert(error);
+    }
+  }
+
+  const handleDelete = (id) => {
+    const confirm = window.confirm("Bạn muốn xóa hàng này ?", id)
+    if (confirm) {
+      BookService.deleteAuthorById(id);
+      setAuthors(authors.filter(item => item.id != id));
+    }
+  }
+
+  const handleUpdate = (oldData) => {
+    setFormData(oldData)
+    handleClickOpen()
+  }
 
   useEffect(() => {
     BookService.getAllAuthor()
@@ -41,28 +75,23 @@ const DatatableAuthor = () => {
       .catch((err) => console.log(err));
   }, []);
 
-  const userColumns = [
+  const authorColumns = [
     { field: "id", headerName: "ID", flex: 0.5 },
     {
-      field: "companyName",
-      headerName: "Company Name",
-      flex:3
-    },
-    {
       field: "firstName",
-      headerName: "First Name",
-      flex: 2
+      headerName: "Họ",
+      flex: 2.5
     },
     {
       field: "lastName",
-      headerName: "Last Name",
+      headerName: "Tên",
+      flex: 2.5
+    },
+    {
+      field: "companyName",
+      headerName: "Tên công ty",
       flex: 7.5
     },
-  ];
-
-  
-
-  const actionColumn = [
     {
       field: "action",
       headerName: "Hành động",
@@ -70,15 +99,16 @@ const DatatableAuthor = () => {
       renderCell: (params) => {
         return (
           <div className="cellAction">
-            <div className="viewButton" onClick={() => handleModalView(params.row.id)}>
-              View
-            </div>
-
+            <div
+              className="viewButton"
+              onClick={() => handleUpdate(params.row)}
+            >
+              Xem</div>
             <div
               className="deleteButton"
               onClick={() => handleDelete(params.row.id)}
             >
-              Delete
+              Xóa
             </div>
           </div>
         );
@@ -86,20 +116,34 @@ const DatatableAuthor = () => {
     },
   ];
 
+
+
   return (
+
     <div className="datatable">
+      {
+        <AuthorModal
+          open={open}
+          handleClose={handleClose}
+          data={formData}
+          onChange={onChange}
+          handleFormSubmit={handleFormSubmit}
+        />
+      }
+
       <div className="datatableTitle">
-        Thêm tác giả
-        <button className="link" onClick={handleModalAdd}>
-          Thêm 
+        Quản lí tác giả
+        <button
+          className="link"
+          onClick={handleClickOpen}
+        >
+          Thêm
         </button>
-        {openmodaladd && <NewAuthor closeModal={setOpenmodalAdd} />}
       </div>
-      {openmodalupdate && <DetailAuthor closeModal={setOpenmodalUpdate} />}
       <DataGrid
         className="datagrid"
         rows={authors}
-        columns={userColumns.concat(actionColumn)}
+        columns={authorColumns}
         pageSize={10}
         rowsPerPageOptions={[10]}
         disableSelectionOnClick
